Improves Layout of TestAnalysisVSPAERO Script and Adds 2 Tests
This reformats the layout of the existing TestAnalysisVSPAERO
script by adding a main function to control the execution of each
individual test. Two additional tests are added, which are similar
to TestVSPAeroSharpTrailingEdge and TestVSPAeroBluntTrailingEdge
in APITestSuiteVSPAERO. These tests are used to demonstrate
creating more complex geometries in a script, performing analyses,
and displaying results. Since both tests use a symmetric airfoil with
no angle between angle of attack and alpha, the coefficient of lift
is expected to be 0 for each.