Adds MeshCommonSettings to geom_core Directory
This creates MeshCommonSettings.cpp and MeshCommonSettings.h and adds
them to the CMakeList. The MeshCommonSettings class is derived from
Parm Container, and will serve as the base class for CFDMeshSettings
and StructSettings. This allows the MeshCommonSettings pointer to
get different values for the same parm from either CFDMeshSettings or
StructSettings, depending on which analysis is running.