Updates CfdMeshSettings and StructSettings with MeshCommonSettings
This updates both StructSettings and CfdMeshSettings to be derived
from MeshCommonSettings, which is derived from ParmContainer
and contains all of the parms and functions common between the
two settings classes. This allows for a pointer to MeshCommonSettings
to be called in CfdMeshMgr to obtain the correct settings according to
whether a Cfd or Fea mesh is running. Both classes are initialized with
their own copies of the MeshCommonSettings parms.