Misc. DXF Cleanup, Renaming, and Reorganization
This improves the naming of DXF variables and parms to add clarity in
preparation of adding similar vairables and parms for SVG export. A NOSA
header is added to DXFUtil. Various reoragnization to alphabetize. Fixes
the default rotation for 4 view DXF exports. Adds parm description and
comments. Removes saving of parm values in DXFOptionsScreen, since
parm values are automatically saved.