Improves. Renames, and Reformats DXFUtil Functions
First, DXFManipulate and DXFShift are renamed in expectation of their
use with SVG exports. Both functions are modified to improve the
alignment of exports, and to rotate and shift feature lines correctly.
Fprintf statements are improved by moving constants to the
format string input. The WriteDXFHeader function is condensed with
a switch/case and the current VSP Version is added. In
FeatureLinesManipulate a check for no view is added and the
DEG_2_RAD enum is implemented.