Adds 2 Trailing Edge Tests to APITestSuiteVSPAERO
The two tests are similar, but are slightly different in the
defenition of the trailing edge of the wing geometry. Both begin by
creating the test wing with a symmetric airfoil, one with a sharp
trailing edge and the other with a blunt trailing edge. VSPAero Compute
Geometry analysis is executed to create vortex lattice degengeom file.
Then, VSPAEROSinglePoint analysis is executed with the angle between
angle of attack and alpha set to 0. The resulting coefficient of
lift is compared to the expected value of 0.