Adds TestVSPAeroSupersonicDeltaWing
This adds an additional test to APITestSuiteVSPAERO to demonstrate a
more unique test case, since most of the tests in the test suite use
the same geometry. The delta wing geometry is created, the vortex
lattice degen geom file is created and saved, and VSPAEROComputeGeometry
is executed with the results displayed.