VSPAERO improves control grouping robustness in CreateInputFile
This commit groups control surfaces and writes successful groupings to a
string buffer before writing them to the *.vspaero input file.  The new
grouping algorithm adds a check for the parent component name prior to
the check for the control surface name.  This allows the user to have
fewer naming restrictions on unique names.