Adds WriteProjectionLinesDXF to Geom
This function is used to get the projection lines for
each view to be exported. The projection lines are
then shifted and manipulated in the same way the
feature lines are to ensure proper alignment. This
function is only called if the flag to export geom
projection lines is true.