Adds WriteFeatureLinesSVG to Geom
This function is similar to WriteFeatureLinesDXF, but unique
to the XML-based SVG file format. This function begins by
looping over each surface of the geom, getting all of the
feature lines, and shifting them to the orgin. The feature
lines are then written to the SVG file, taking account of the
views and rotations indicated.