Make ConformalGeom work with non-fp controlled skinning.
Before, AdjustShape assumed all ribs set right_fp and left_fp values.  It
would scale those values to optimize the skinning accuracy.

If those values were not used, it would cause the ReSkin to crash.  This
change makes sure to only change _fp values that are in-use.  It also adds
a fast-path exit for the case where no _fp values are used.