Adds m_VehProjectVec3d to Vehicle
This adds a vector of projection lines, in the X, Y, and Z
directions. Functions are included to get and set indexes
of this vector. m_VehProjectVec3d will be empty, other
than for when an export is in progress and the flag to
export vehicle projection lines is indicated.