Adds Inputs to ProjectionMgr Functions
First, MeshToPathsVec and PathsToPolyVec are updated with two new
inputs: keepdir1 and keepdir2. These correspond to which axes to keep,
and are set to 1 and 2 in the header file to maintain the standard of
projecting along the x axis. An input flag is added to
BuildToFromClipper to translate to the bounding box maximum (the
default setting) or to the center. These functions will be called in
ExportProjectionLines.