Adds XSecFeatureLine Forcing Option
A flag is added to Geom to force feature lines at xsecs, where
feature lines normally would not be drawn. This flag is added
as an input to BuildFeatureLines, but is set to false in the
header file. If xsecs are to be included in the export, this flag
will be set to true, the geom will be updated, and the flag will
be reset to false after the export is complete.