Improves WriteDXFFile by Implementing New DXF Options
This improves WriteDXFFile by implementing the
additions of geom and total projection line exports. The
function begins by increasing the tesselation of each geom
in the indicated set if projections will be exported. For geoms
with multiple sections, the U tesstelation is adjusted for each
section. The header of the DXF file is written and the feature
lines of each geom are written. The total vehicle projection
lines are written if indicated. The tesselation values for each
geom are restored if changed. Last, the closer is written.