Adds WriteSVGFile to Vehicle
This function is similar to WriteDXFFile, but unique to the
XML-based SVG file format. This function begins by
increasing the tesselation of each geom in the indicated set
if projections will be exported. The header of the SVG file is
written and the feature lines of each geom are written. The
total vehicle projection lines are written if indicated. The
tesselation values for each geom are restored if changed.
Last, a scale bar is created if the flag is true.