Adds Two MeshGeom Functions for Writing Facet Nodes and TriParts
The first function lists all of the XYZ facet vertices (nodes). The
second determines the number of facets for each part and assignes three
nodes to each facet. The function loops trough a vector of tags to
separate components. Other information written to each facet includes
the component ID, material ID, and running number of facets. Material
ID is set to the default of 0 at this point.