Adds New Options to DXFOptionsScreen
This adds the GUI elements for specifying to export geom projections,
total vehicle projections, and cross sections, along with the options to
color exported feature lines, append the geom ID to each layer, and to
modify the tesselation to improve projection lines resolution. The
layout of the screen is adjusted to accomodate these additions.