BUG FIX: Sets m_MeshInProgress Flag Before Starting Thread Processes
This fixes a bug in the output display for a CfdMesh. In some cases no
text would be added to the display. This was caused by the
m_MonitorProcess staring prior to m_MeshInProgress being set in
CfdMeshMgr, allowing m_MonitorProcess to terminate prematurely. Setting
the flag in CfdMeshScreen before begining each thread process ensures
this will not occur.