Adds the Class FeaStructure
This class is the base class that defines an FeaStructure. Every
FeaStructure has a unique ID, a parent geom ID, a vector of FeaParts,
and a vector of subsurfaces. Functions are added to manage these vectors
by adding, deleting, and validating their indexes. XML support for
FeaStructures is also included, along with a GenerateID wrapper
function.