Improves GetScale Function and Updates m_SVGSet
In Export Screen, m_SelectedSetIndex is used to update m_SVGSet, which
will be used to approximate the scale bar based on the geometries in the
export set. The rounding routine in GetScale is improved and a check for
no geoms in the set is added. Last, an unused input to
ShowSVGOptionsScreen is removed.