Adds OrientStructure Options to StructScreen
This adds a function, OrientStructure, that shows only the parent geom
of the selected structure, sets only feature lines to be drawn, sets the
view direction, and fits and centers the view. A front, side, and top
orientation setting are provided for non-wing geom types, which are
oriented with a top or side view. Note: this commit may be replaced
by a degen-geom approach to 2D structural layouts.