Adds a Function to Define DegenGeom DrawObj Vectors
This adds four vectors of DrawObj: DegenSurfaces, DegenPlates, Degen-
CamberedPlates, and DegenSubSurfaces. UpdateDegenDrawObj is
created to calculate the degenerate representation of the geom. The
results are used to update the plate and surface DrawObj vectors.
Only the point (x) and normal (n_vec) data is used to define the quads
for the plate or surface. The cambered plate is calculated using the
 zcamber vector. The DegenSubsurfs are recreated by interpolating the
xyz values of uw points on the Geom surface from the DegenSubSurface.