Updates LoadDrawObj with DegenGeom DrawObj Vectors
In reference to the value of the display type enum the Degen Surface,
Plate, or Cambered Plate are loaded with the other DrawObj for the
Geom. The Bezier wire mesh, feature lines, and subsurface DrawObj
vectors are only loaded if the DISPLAY_BEZIER enum is indicated. The
DegenGeom material (for OpenGL rendering) is set to the
default material. In GeomXSec LoadDrawObjs, the xsecs are only
displayed if the Bezier surface is displayed.