Adds Display Type Choice to ManageGeomScreen
This adds a choice dropdown for specification of the display
type for a Geom or group of Geoms. The available display
types are Surface (Bezier), Degen Surface, Degen Plate, and
Degen Cambered Plate. Functions are added to update and
set the display choice, and the screen size is adjusted
accordingly.