Updates Treatment of Symmetric FeaRibs and FeaParts
This commit is similar to the symmetric treatment updates that were made
to FeaFullDepth and FeaStiffenerPlane FeaParts. Instead of defining
symmetric copies of these FeaParts by looping over each symmetric
surface of the parent geom, a single FeaPart is definied from the
indicated main surface of the parent geom, duplicated for each
symmetric copy, and transformed accordingly by calling
CalculateSymmetricTransform().