Improves Orientation Functions in StructScreen
This adds a specific function, OrientWing, for specifying the view type
for wing structures, reducing the options of "top" or "side" to a single
button. The DegenGeom previews are shown according to the
specified view. The orientation GUI groups are fixed to show or hide
correctly, additional pointer protection is added, and the currently
selected structure is used to identify the current geom.