Removes Generated FeaStructID and Adds GetFeaStructIndex Function
The FeaStructID was only being used sparingly, and was mixed with code
that only used the FeaStructure index. To save memory and avoid multiple
methods of referencing a FeaStructure, the generated FeaStructID was
removed. In StructScreen, the FeaStructID use is replaced by a call
to GetFeaStructIndex, which has been added to StructureMgr.