Removes Tracking of FeaPart Types in Surf and FeaElement
This sets up a change in FeaMeshMgr to no longer distinguish different
FeaPart types. Surfaces in FeaMeshMgr will be given a more general
treatment, allowing for smoother integration of new FeaPart types in the
future. All of the information needed for FeaParts will carry though the
FeaMeshMgr with the FeaPart index alone. In FeaElement, the FeaProperty
index is kept for use in WriteNASTRAN and ComputeMass.