Adds 2 Functions and Removes 1 from FeaStructure
First, the version of GetFeaPart that takes the FeaPartID as input is
removed, since this function is included in StructureMgr and the
identification of FeaPart by index is preferred in FeaStructure.
GetFeaPartName is added to return the name of an FeaPart given the
index. Similarly, GetFeaPropertyIndex is added to return the
FeaPropertyIndex given an FeaPart index.