Removes Tracking of FeaProperty Index and Modifies Printed Indexes
This removes the setting of FeaProperty index to each FeaElement in
FeaMeshMgr. To maintain the correct assignment of FeaProperties to
FeaElements, the property index is added as an input to WriteNASTRAN and
ComputeMass in FeaElement. This additional input is identified in
FeaMeshMgr using the FeaPart index and the GetFeaPropertyIndex function.
Additionally, the index for FeaProperties and FeaMaterials is increased
by 1 wherever it is written to file, avoiding a 0 index.