Adds Access to *.msh Export for Fea Meshes
First, the GMSH_FEA_NAME enum is defined in APIDefines, and is then used
to add an additional m_ExportFileFlag and file name suffix in
StructSettings. The previously defined WriteGmsh function is called in
ExportFeaMesh. Last, the GUI elements to set the *.msh file name,
destination, and export flag are added to the StructScreen.