Adds FeaBeam Intersection Caps to BuildFeaMesh
This updates BuildFeaMesh to define FeaBeam elements at the
intersections between an FeaPart and skin surface. Checks are
included for the FeaPart cap intersection flag and for one of the
surfaces at the  intersection being an FeaPart while the other is a
skin surface. The normal vector for the FeaBeam is defined from
the skin surface normal. Also, the direction of the intersection
points about the center of the surface is checked, and the points
are reversed if negative.