Adds Intersection Cap Parms and Functions
A BoolParm flag is added to FeaPart to indicate whether or not
beam elements should be defined at the intersections of the
FeaPart and geom surface (skin). An additional BoolParm is
added to indicate if the caps will be drawn. A FeaProperty
index is addded for caps, and a function is added to get the
FeaProperty index from an FeaPart index. XML encode and
decode support is also updated.