Adds GUI Elements for Intersection Caps and Cap Properties
This updates StructScreen with GUI elements for setting the flag
to add caps at FeaPart intersections, and to define the cap
properties. Cap elements are added to the Draw Part Browser.
Two vectors of names in the browser and the index of their
FeaPart are added to maintain the drawing parm settings.