Adds Beam Element Normal Drawing Option
First, a new vector of DrawObjs is added to FeaMeshMgr and
is defined as a line from the end of an FeaBeam to a point
located on the normal vector for the FeaBeam. A BoolParm
is added to StructSettings to indicate if the normal vectors
should be drawn. StructScreen is updated to support the
new option.