Reorganizes Element and Cap Drawing Parms and Settings
This removes drawing parms for FeaElements and intersection caps in
FeaStructure and replaces them with vectors of bools in FeaMeshMgr. At
the end of FEA Mesh generation, the data for the vectors of DrawObjs is
initialized. The GUI elements in StructScreen are used to update these
vectors, and the Draw Part Browser is updated with the vectors in
FeaMeshMgr. All of this is done to avoid invalid pointers when an
FeaStructure is deleted or changed.