Improve CustomGeom name matching
Add concept of display name to show on-screen when internal names have been
unique-ified to resolve collisions.

Use module script name to match which GUI is displayed instead of the
basic type name.  The type name is set correctly when scripts are read
from CustomScripts, but not when read from a *.vsp3 file.

The module name is used to identify which script should be called, so it
should be used for consistency no matter what.