Removes FeaStiffenerPlane and FeaStiffenerSubSurf
The FeaStiffenerPlane class is removed in preparation of replacing
its functionality with a flag for FeaFullDepth parts to include
interior tris. The FeaStiffenerSubSurf class had not been fully
defined. It will be developed using the existing methods for
line-type FeaSubSurfaces, with options added to only include the
intersection beam elements.