Modifies BuildSubSurfIntChains and Adds ISegChain SubSurface Index
First, a subsurface index is added to ISegChain to identify which
subsurface in m_FeaSubSurfVec the intersection chain belongs to.
BuildSubSurfIntChains is duplicated from CfdMesh in FeaMesh, but
is modified to get only the FeaSubSurfaces, prepare them to be split,
and adds the subsurface index to the intersection chain if beam
elements are to be defined along the intersection curve.