Adds Subsurface Intersection Cap Beam Element Support
First, a BoolParm flag is added to SubSurface to identify if beam
elements will be defined for FeaSubSurface intersection lines. In
addition, an FeaProperty index is included with methods to get
and set the index. XML support for the index is also added. In
FeaMeshMgr, BuildFeaMesh is modified to add a check for tris
with multiple tags, indicating a subsurface. m_ISegChainList is
provided an additional check for intersections with the same
component ID, which is expected for a subsurface intersection.