Fixes Incorrectly Flipped Normal Direction
Previously, the flip normal flag for the surface was used to set the
flip normal flag for the degenerate geom DrawObj. The result of this
approach saw incorrecly flipped normals, such as for the upper
surface of one side of a symmetric wing. Flipping the normal vector
directly avoids this bug.