Adds DisplayDegenCamber Preview to VSPAEROScreen
In VSPAEROMgr, a BoolParm flag is added to indicate drawing of
DegenGeom Cambered Plate previews for VLM analysis. In
VSPAEROScreen, the button to Compute Geometry is replaced
by a toggle to preview VLM DegenGeoms. ComputeGeometry is
called int the GUI prior to executing the solver. A function is added
to set the display settings to Degen Camber for each geom in the
set. An additional function restores the geoms to a Bezier wireframe
display.