BUGFIX: Fixes Loss of SubSurface Index Tracking in IntersectSplitChains
The SubSurface index for the split intersection chains was being lost
when new intersection chains were created and added to m_ISegChainList.
The subsurface index for the new chain is set equal to the index for
the original chain it was created from. This is also done in
SplitBorderCurves for consistency, but is not used.