Modifies Tag Identification in BuildFeaMesh
Instead of calling SubSurfaceMgr.GetTag(), the second index of the tag
is used if the number of tags for the tri is equal to 2. This is done
to correctly set the FeaSubSurface for FeaTris when "None" is selected
for the inclusion of FeaSubSurface tris, since they are not added to
m_TagCombos. In addition, more than two tags for a tri indicates two
or more subsurfaces overlapping, and is set as a TODO item for now.