BUGFIX: Moves DegenSubSurface Updates from UpdateDrawObj to ParmChanged
A bug was identified when loading a model with two control subsurfaces
on a wing. degenGeom.addDegenSubSurf was being called prior to the
second control subsurface being updated (indexing error for m_LVec).
Moving the call to UpdateDegenDrawObjs from UpdateDrawObj to ParmChanged
acomplishes the same functionality with no risk to decoded XML data.