Adds SetFixPoint to Mesh
This function iterates over the nodeList to find the closest node
to the given FeaFixedPoint 3D coordinate. The closest point is then
moved to the closest UW point on the surface, using the 3D and UW
coordinates of the FeaFixedPoint. The minimum Grid Density length
is used as a tolerance to check the edges connected to the node,
which are then split if too long. If no closest node can be found,
the edges are split and the function is called again. The function
will return false if no closest node can be found and no edges can
be split.