Fully Defines FeaFixPoints
FeaFixPoints are defined by a parent FeaPartID, which is used to
identify the parent surfaces the FeaFixPoint is attached to.
In IdentifySplitSurfIndex, the parent surfaces are split and the index
of the split surface where the FeaFixPoint lies is saved. Functions are
added to get the 3D coordinates of the FeaFixPoint on each parent
surface, and the UW coordinate. Last, LoadDrawObj is updated to
support the changes.