Updates Tagging and Adds FixPointFlag to FeaNode
This adds a flag to FeaNodes to indicate if it is an FeaFixPoint.
In TagFeaNodes, the 3D coordinates of each FeaNode in
m_FeaNodeVec are compared to each point in m_FixPntVec.
If the points are equal, m_FixedPointFlag is set to true and the
FeaNode is tagged with the FeaPart index. In BuildFeaMesh,
FeaFixPoints are not included in FeaSubSurface indexing.