Updates and Reorganizes LoadDrawObj
First, m_FixPointFeaPartFlagVec is added and defined in
UpdateDrawObjData to indicate if an FeaPart is of
FeaFixPoint type. in LoadDrawObj, this flag is used to identify
if m_FeaElementDO will be defined as points or tris. Last, the
two for loops that iterate over the number of FeaParts are
combined.