Adds SetFixPointBorderNodes
SetFixPointBorderNodes is added to iterate over the vector
of m_Surf indexes for each FeaFixPoint. If the border flag for
the FeaFixPoint is true and there are two surface indexes for
it, a split is added between the two closest points on the
border curve. Nothing is done if there are more than two
surface indexes for the FeaFixPoint, since a node will always
be placed at the intersection of two or more border curves.