Adds Support for FeaFixedPoints on Symmetric Surfaces
In FeaStructure, IdentifySplitSurfIndex is updated to use a
vector < vector < int > > m_SplitSurfIndex, where the vectors
size corresponds to the number of parent surfaces. The
indexes for the symmetric surfaces will be offset from the
first parent surface by the number of split surfaces.
FeaMeshMgr is udated to support these changes.