Initializes Data and Improves Transfer From FeaStructure to FeaMeshMgr
Information from FeaStructure is transfered and saved in FeaMeshMgr
at the start of the mesh process. This is done in an effort to
have FeaMeshMgr be a stand-alone program that does not need to rely
on files in geom_core during the mesh process. In addition, variables in
FeaStructure and FeaMeshMgr are initialized in the constructor. Extra
pointer protection is included during the data transfer process. Note:
SubSurface data and functions have not been completely transfered at
this time.