Adds Full Support for SS_Line Degen SubSurface Previews
DegenSubSurfaces that are type SS_Line are only defined by their start
and end uw coordinates. Previously, this would result in SS_Line Degen-
SubSurfaces to only be drawn as a line or point (for closed U/W
surfaces). A check is added for SS_Line typeIDs, and the DegenSubSurface
is identified as constant U or W. Linear interpolation is then used to
find all of the points on the DegenSurface that correspond to the
constant U/W value.