Creates SimpleSubSurface in cfd_mesh Directory
This creates the SimpleSubSurface .cpp and .h files and adds them to the
CMakeList. The SimpleSubSurface class is defined as a simplified version
of the SubSurface class. The functions used in cfd_mesh for SubSurfaces
are copied here and slightly modified for use with primitive data types,
which are set from SubSurface Parms in the CopyFrom function. This is
done to limit reliance on SubSurface pointers in cfd_mesh and data
transfer between the directory and geom_core.