Adds Tracking of SimpleSubSurfaces and Data Transfer Functions
First, a vector of SimpleSubSurfs is added as m_SimpleSubSurfaceVec. It
is defined in TransferSubSurfaceData for both CfdMesh and FeaMesh, so
that the SimpleSubSurfaces are copied from the appropriate SubSurfaces.
In FeaMeshMgr, the SubSurface data vectors are removed, since their
information is contained in the SimpleSubSurface class. The calls to
the old SubSurface data vectors is updated accordingly. Last, in
CfdMesh, functions that used SubSurfaces are updated to use the
SimpleSubSurfaces and work for both Cfd and Fea meshes.