Adds SimpleSubSurface Management Functions to CfdMesh
Two of the functions, GetSimpSubSurfs and SetSimpSurfTags, are virtually
identical to functions used with normal SubSurfaces in SubSurfaceMgr,
but are tailored to use SimpleSubSurfaces. SubTag in Surf is copied and
modified with SimpleSubSurfaces in CfdMeshMgr. It has been moved to
avoid data loss from leaving the scope of CfdMeshMgr and to avoid
needing SimpleSubSurface pointers.