Creates the SimpleFeaProperty and SimpleFeaMaterial Classes
These classes are added to FeaElement to save FeaProperty and
FeaMaterial Parm data in more primative data types. The functions for
writing these classes to file are also copied from FeaStructure with
slight modifications. This is done in an effort to redude reliance on
geom_core and FeaPropertry/FeaMaterial pointers in FeaMeshMgr.