Adds Data Transfer and Tracking of SimpleFeaProperties/Materials
In FeaMeshMgr, two vectors of SimpleFeaProperties and SimpleFeaMaterials
are added with functions to get the vectors. The ComputeMass functions
in FeaElement are then updated to use these vectors instead of
FeaProperty/Material pointers. The transfer of FeaProperty and
FeaMaterial data is added to TransferFeaData in FeaMeshMgr. Calls to
StructureMgr or other uses of FeaProperties/Materials are updated to
use SimpleFeaProperties/Materials.