Improves FeaFixPoint Surface and Intersection Identification
This improves the existing method of checking for FeaFixPoints on
intersection curves and includes support of FeaFixPoints on
SubSurface intersection curves. In Mesh, a change is made to
SetFixPoint so that the closest node is not moved until all edge
splitting is complete. In CfdMeshMgr, CheckFixPointIntersects is
moved after SplitBorderCurves is called. Output messages are
added to identify successful setting of FeaFixPoints.