Implement ancestor grouping for better total Swet calcs
This will implement grouping of certain geometries that are better
suited to have their wetted areas included in a the wetted area of
a parent geometry causing the grouped geometry to take the properties
of its' parent geometry such as form factor equation choice and
reference length. (e.g. a gear pod to a fuselage)