Makes StructSettings and GridDensity Unique to Each FeaStructure
FeaGridDensity and StructSettings are added to the FeaStructure
class and removed from Vehicle. Updates are made to get the
grid density and settings pointers from FeaStructure. Two
functions are added to StructureMgr to reset the file export
names and set all drawing flags to false. XML support is also
modified.