Fixes Incorrect Indexing of FeaFixPoints for Half Meshes
This adds two functions to FeaFixPoint, LessThanY and PlaneAtYZero,
which are almost identical to the functions of the same name in
SurfCore. A flag is added to track the value of the Half Mesh Flag in
MeshCommonSettings, updated in UpdateFeaParts. This flag and the two
added functions are used to check if the split surface index should
be included for the FeaFixPoint.