Replaces GridDensity with SimpleGridDensity in Mesh Support Files
All calls to the GridDensity class in the geom_core directory are
replaced with calls to SimpleGridDensity in the cfd_mesh
directory in ISegChain, SCurve, Mesh, and Surf. This is a part
of the transition of mesh settings from Parm values to
more primitive data types, saved in the SimpleMeshSettings
classes at the beginning of a mesh process.