Updates Mesh Settings Pointers in CfdMeshScreen and the API
This updates the calls to CfdMeshMgr mesh setting and grid density
pointers with pointers to the classes maintained in Vehicle. This
completes the transition of mesh Parms and pointers in geom_core
to SimpleMeshSettings in cfd_mesh, greatly reducing the number
of times data is transfered between the two directories.