CheckSubSurfBorderIntersect Bug Fixes and Improvements
Considerations for MagicV Parms fixes a bug for identifying SubSurface
borders on wing leading or trailing edges. Adding checks for ClosedU or
ClosedW surfaces fixes a bug with wing tips being incorrectly set as
SubSurface borders. Last, additional checks for valid UW points are
added to ensure running in debug mode is successful and no error or
warning messages are printed to the console.