Adds OnValidPatch Function to VspSurf
This function is used to check if a given UW point is on
a valid patch, which is one that will be included when
the surface is split prior to a mesh process. The function
checks that m_MagicVParm is true before splitting the
surface and calling CheckValidPatch. The UW limits of the
patch are used to validate the input UW point.