Adds Option to Set Ribs Perpendicular to Spars
First, the IntParm flag to indicate which edge a rib is perpendicular
to is changed to an int, where an index greater than 2 indicates the
rib is perpendicular to a spar. In StructScreen, the perpendicular
edge choice for ribs is reorganized to include all spars in the
UpdatePerpendicularRibChoice function. Tracking of perpendicular
spars is done by FeaPart ID.