Improves Resizing of FeaSlices
The routine used to define the planar surface of FeaSlices is
modified to take into consideration rotational changes of the
FeaSlice. The FeaSlice is resized to intersect the bounding box of
of the surface. Prior to this change, rotations would cause
FeaSlices to only partially pass through the parent geometry.