Combines FeaSubSurfaces with FeaParts Tab
This merges the two tabs and browsers for FeaParts and
FeaSubSurfaces, along with the associated update functions.
The selected index in the FeaPart Browser will be an FeaPart if
the index is less than the total number of FeaParts, or an
FeaSubSurface if the index is greater than or equal to the
number of FeaParts. Also, RecolorFeaSubSurfaces is modified
to highlight a vector of SubSurfaces.