Adds FeaPart Browser Reordering
This is similar to what was previously made available for Fea
SubSurfaces, but has been added to the FeaPart Browser to work for all
FeaParts and SubSurfaces. Since FeaParts and FeaSubSurfaces are stored
in separate vectors, reordering is only availble for within the limits
of the vector (FeaParts and FeaSubSurfaces remain grouped together).
The ReorderFeaParts function is added to FeaStructure to support the
addition.