Modifies Tracking of FeaPart Display from FeaStructure to FeaParts
This is done to allow for inidivual FeaParts or groups of FeaParts to
be displayed or hidden. This applies to FEA SubSurfaces as well. A
function is added to FeaStructure to set all FeaPart display parms.
In GuiDeviceCallBack, the display settings of multiple selected
FeaParts can all be set to the same value.