Adds Support for FeaProperty Changes for Multiple FeaParts
If more than one FeaPart is selected, the FeaProperty index
of each selected part is checked. If they are all the same,
the Property choice shows the FeaProperty index of the first
selected part. If there are not all the same, "Multiple" is
displayed. All selected FeaParts can be changed to the same
FeaProperty if the property index is compatible with the
element type.