Improves and Reorganizes StructScreen and FeaPartEditScreen
In FeaPartEditScreen, a call is added to update the FeaStructure
if an FEA Mesh is not currently running. Changes to GUI elements in
the FeaPartEditScreen now update the FeaParts in real-time. Next,
the StructScreen Part tab is reorganized slightly, additional
considerations are added for FeaFixPoints, FeaParts are selected after
being added, comments are added, and an indexing error is corrected.