Adds Support for FeaPointMasses to FeaMeshMgr
This adds tracking of m_FixPointMassFlag and m_FixPointMass
to FeaMeshMgr through the data transfer in AddStructureParts.
Tagging of FeaFixPoints is improved so that only those that make up
triangles (which have already been tagged) are tagged as FeaFixPoints.
If the FeaFixPoint is indicated to be a point mass, an FeaPointMass
is created. Last, FeaPointMasses are added to the export files and
the export functions are reorganized slightly.