Adds Support for Identifying Multiple Splits on SubSurface Edges
In CheckSubSurfBorderIntersect, all of the split points are collected
from m_ISegChainList. Next, the split points are compared with the
subsurface edges for each intersection chain to determine if the
subsurface edge has been split. Checks are also included to ensure the
split point lies on the subsurface edge. All together, this commit
provides correct beam element identification for overlapping subsurface
edges, fixed points on subsurface edges, and similar cases.