New Unit Conventions for FEA Structures
Instead of allowing the user to specify length, mass, and
temperature units, a series of unit systems are introduced
which guarantee unit consistency in the output file formats.
Associated parms are automatically converted when the
unit system is changed. All functions and parms associated
with the previous method for specifying units for structures
are removed.