SimpleSubSurface Cleanup and Improvements
Instead of copying the UpdatePolygonPoints function from SubSurface
to SimpleSubSurface, the polygon points are updated in the CopyFrom
function. This removes any additional need to update the polygon points,
since they will not change during the FEA mesh process. In addition,
changes made to UpdatePolygonPoints in SubSurface no longer need
to be duplicated in SimpleSubSurface.