Sets m_TestType to Inside when Included Elements are only Beam Elements
This forces the interior triangles of the SubSurface to be removed
at the end of the FEA mesh process and only the intersection curves
seved as beam elements. If the triangles on the outside of the
SubSurface are desired to be removed, m_RemoveSkinTrisFlag in FeaSkin
can be set appropriately. In addition, the SSLine parm updates are
reorganized for consistency, but m_TestType is not set.