Changes Name of FetchXFerSurf Used in FeaStructure
Although similar to FetchXFerSurf in the fact that the function
breaks a VspSurf into XFerSurfs, it was not appropriate to have the
same name for another function that does an additional check
for if an input UW point is on a valid patch. Therefore, the name
of the function is changed to CheckandFetchXFerSurf.