Reformats FeaPart Updates and Definition of Symmetric Copies
First, CalculateSymmetricTransform is removed and replaced by a call to
Geom:GetFeaTransMatVec in the newly added function UpdateSymmParts.
UpdateSymmetricSurfs is renamed to UpdateSymmIndex. In FeaStructure::
UpdateFeaParts, the symmetric indexes are first updated, the FeaPart
is updated (primary FeaPart surface is defined), and then the symmetric
FeaPart surfaces are created. This is all done in an effort to reduce
reusing code from Geom and reduce redundant FeaPart updates.