Adds One Roation and Improves Existing FeaSlice Rotations
Rotations of FeaSlices are specified by their rotational axis: X, Y, or
Z. The parm names and labels are updated to reflect this. In addition,
rotations are limited to a single axis at a time, since multiple axes of
rotation would allow FeaSlices to extend beyond the bounding box
for the parent geometry, potentially intersecting other FeaParts.