Adds Options for Spar Position and Trimming to a Wing Section
First, the parameterization of FeaSpar location is updated to be
defined from either percent of chord or average chord length. Next,
an option is added to trim FeaSpars to a wing section, in which the
current wing section is saved as an IntParm. GUI elements are added
to the FeaPartEditScreen to support these added features.