Changes FeaPart Center Location from Decimal to Percent
Instead of being represented as a decimal, the center location for
FeaParts is changed to a percent value if percent parameterization is
indicated. In ComputePlanarSlice, the percent value is converted back to
a decimal. In addition, FeaSpars are adjusted to range from the start to
end of the average chord distance, and no longer loop back to the
starting point.