Creates the FeaBulkhead Class in FeaStructure
This commit adds the class FeaBulkhead, derived from FeaPart.
FeaBulkhead is based off EllipsoidGeom, but is defined as a
body of revolution of a half curve in BuildBulkheadSurf. 9
Parms are used with FeaBulkhead: x,y, z translation, x, y, z
rotation, and a, b, c radius length. The initial center of the
FeaBulkhead is set to the nose (uw = 0,0) of the parent geom.