Adds and Defines the FeaRibArray Class
This adds a new class, FeaRibArray, derived from FeaPart. In
CreateFeaRibArray, the number of ribs and spacing between them
is defined from the parms m_StartLocation and m_RibSpacing.
AddFeaRib is then called to add FeaRibs to the array and assign
data to each rib, such as FeaProperty index and perpendicular
edge index.