Adds GetRibTotalRotation and Removes m_PerpendicularEdgeIndex
m_PerpendicularEdgeIndex was identified as unnecessary, since a
string comparison with m_PerpendicularEdgeID can be used to
determine if an FeaRib shall be perpendicular to the leading edge,
trailing edge, or an FeaSpar. GetRibTotalRotation is added to FeaPart
to calculate the rotation of an FeaRib. This is done in preparation
of use by FeaRibArray.